+++ PubMatic/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 4.7.0.0
+* Certified with PubMatic SDK 4.7.0.
+* Update ad display failed error code.
+
 ## 4.6.0.0
 * Certified with PubMatic SDK 4.6.0.